libraries/test-helper/pyproject.toml
[tool.coverage.run]
branch = true
source = ["test_helper"]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/libraries/test-helper/html' --cov-report xml:'../../coverage/libraries/test-helper/coverage.xml' --html='../../reports/libraries/test-helper/unittests/html/index.html' --junitxml='../../reports/libraries/test-helper/unittests/junit.xml'"

[tool.poetry]
name = "test-helper"
version = "1.0.0"
description = "Palateful Testing Helper Library"
authors = ["Leonid Belyi"]
license = "Proprietary"
readme = "README.md"

[[tool.poetry.packages]]
include = "test_helper"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^8.0"
pytest-cov = "^6.0"
pytest-html = "^4.0"
pytest-asyncio = "^0.25"
factory-boy = "^3.3"
faker = "^37.0"
utils = {path = "../utils", develop = true}

[tool.poetry.group.dev.dependencies]
ruff = "^0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
